added error throwing

+++ server/src/main/java/service/UserService.java
@@ -18,6 +18,9 @@ public class UserService {
     }
 
     public RegisterResult register(RegisterRequest registerRequest) throws Exception {
+        if (registerRequest.username() == null || registerRequest.password() == null){
+            throw new BadRequestException("missing username or password");
+        }
         if (userDAO.getUser(registerRequest.username()) != null){
             throw new AlreadyTakenException("username already taken");
         }
@@ -34,7 +37,11 @@ public class UserService {
         authDAO.createAuth(new AuthData(token, registerRequest.username()));
         return new RegisterResult(registerRequest.username(), token);
     }
+
     public LoginResult login(LoginRequest loginRequest) throws Exception{
+        if (loginRequest.username() == null || loginRequest.password() == null){
+            throw new BadRequestException("missing username or password");
+        }
         if (userDAO.getUser(loginRequest.username()) == null){
             throw new BadRequestException("username doesn't exist");
         }
@@ -48,8 +55,15 @@ public class UserService {
         authDAO.createAuth(new AuthData(token, loginRequest.username()));
         return new LoginResult(loginRequest.username(), token);
     }
+
     public void logout(String authToken) {
+        if (authToken == null){
+            throw new UnauthorizedException("unauthorized");
+        }
         AuthData auth = authDAO.getAuth(authToken);
+        if (auth == null){
+            throw new UnauthorizedException("unauthorized");
+        }
         authDAO.deleteAuth(auth);
     }
 }
